+++ out/utils.js
@@ -1,0 +1,4 @@
+"use strict";
+function getTyped(line, col, starts, ends, allow) {
+}
+//# sourceMappingURL=utils.js.map